+++ src/classes/VSCommands/InsertCustomClassNameCommand.ts
@@ -1,0 +1,19 @@
+import * as vscode from "vscode";
+import { FileReader } from "../Util/FileReader";
+export class InsertCustomClassNameCommand {
+	static insertCustomClassName() {
+		const editor = vscode.window.activeTextEditor;
+
+		if (editor) {
+			const document = editor.document;
+			const classNameDotNotationToInsert = FileReader.getClassNameFromPath(document.uri.fsPath);
+			if (classNameDotNotationToInsert) {
+				editor.edit(editBuilder => {
+					if (editor) {
+						editBuilder.insert(editor.selection.start, classNameDotNotationToInsert);
+					}
+				});
+			}
+		}
+	}
+}